✨ HBJ5P8 code: resolve main hotspot drift
