+++ packages/agentplane/src/backends/task-backend/cloud-backend.ts
@@ -1,9 +1,9 @@
 import path from "node:path";
-import { mkdir, readFile, writeFile } from "node:fs/promises";
 import { randomUUID } from "node:crypto";
 
 import { loadDotEnv } from "../../shared/env.js";
 import { isRecord } from "../../shared/guards.js";
+import { readCloudBackendState, writeCloudBackendState } from "./cloud-backend-state.js";
 import {
   BackendError,
   type TaskBackend,
@@ -136,11 +136,7 @@ export class CloudBackend implements TaskBackend {
     fetchImpl?: typeof fetch;
   }): Promise<CloudBackend> {
     await loadDotEnv(opts.root);
-    return new CloudBackend(opts.settings, {
-      root: opts.root,
-      cache: opts.cache,
-      fetchImpl: opts.fetchImpl,
-    });
+    return new CloudBackend(opts.settings, opts);
   }
 
   async generateTaskId(opts: { length: number; attempts: number }): Promise<string> {
@@ -322,13 +318,13 @@ export class CloudBackend implements TaskBackend {
       }
     }
     if (opts.direction === "pull") {
-      await this.writeState({
+      await writeCloudBackendState(this.statePath, {
         last_checked_at: pull.lastCheckedAt ?? new Date().toISOString(),
       });
       return;
     }
     if (!pull.lastCheckedAt) return;
-    await this.writeState({
+    await writeCloudBackendState(this.statePath, {
       last_checked_at: pull.lastCheckedAt,
     });
   }
@@ -495,7 +491,7 @@ export class CloudBackend implements TaskBackend {
 
   async inspectConfiguration(): Promise<TaskBackendInspectionResult> {
     const missing = this.missingConfigKeys();
-    const state = await this.readState();
+    const state = await readCloudBackendState(this.statePath);
     const syncState =
       missing.length === 0
         ? await this.requestCloudSyncState(this.projectId).catch(() => null)
@@ -553,7 +549,7 @@ export class CloudBackend implements TaskBackend {
   }
 
   private async assertProjectionFreshForLocalMutation(): Promise<void> {
-    const state = await this.readState();
+    const state = await readCloudBackendState(this.statePath);
     if (!isStale(state.last_checked_at, this.staleAfterSeconds)) return;
     throw new BackendError(
       [
@@ -583,27 +579,5 @@ export class CloudBackend implements TaskBackend {
     if (!this.token) missing.push("AGENTPLANE_CLOUD_TOKEN");
     if (!this.projectId) missing.push("AGENTPLANE_CLOUD_PROJECT_ID");
     return missing;
-  }
-
-  private async readState(): Promise<{ last_checked_at: string | null }> {
-    try {
-      const raw = JSON.parse(await readFile(this.statePath, "utf8")) as unknown;
-      if (
-        raw &&
-        typeof raw === "object" &&
-        typeof (raw as { last_checked_at?: unknown }).last_checked_at === "string"
-      ) {
-        return { last_checked_at: (raw as { last_checked_at: string }).last_checked_at };
-      }
-    } catch (err) {
-      const code = (err as { code?: string } | null)?.code;
-      if (code !== "ENOENT") throw err;
-    }
-    return { last_checked_at: null };
-  }
-
-  private async writeState(state: { last_checked_at: string }): Promise<void> {
-    await mkdir(path.dirname(this.statePath), { recursive: true });
-    await writeFile(this.statePath, `${JSON.stringify(state, null, 2)}\n`, "utf8");
   }
 }
