+++ scripts/oversized-test-baseline.json
@@ -3,7 +3,7 @@
   "threshold_lines": 1000,
   "budgets": {
     "max_entries": 12,
-    "max_total_lines": 13575,
+    "max_total_lines": 13649,
     "max_file_lines": 1300
   },
   "entries": [
@@ -37,7 +37,7 @@
     },
     {
       "file": "packages/agentplane/src/backends/task-backend.cloud.test.ts",
-      "lines": 1028
+      "lines": 1075
     },
     {
       "file": "packages/agentplane/src/commands/doctor.command.open-pr.test.ts",
@@ -45,7 +45,7 @@
     },
     {
       "file": "packages/agentplane/src/commands/doctor.command.runtime.test.ts",
-      "lines": 1011
+      "lines": 1040
     },
     {
       "file": "packages/agentplane/src/commands/shared/task-store.test.ts",
